Preserve the back stack when navigating to MapFragment
This change fixes an annoying user experience issue - after opening
MapFragment from a secondary screen it was not possible to get back
to that secondary screen by pressing back button.

The reason for this was that the back stack was popped up to the
MapFragment destination which was assumed to be at the bottom of
the stack.

This commit changes this behavior and navigates to a new instance
of MapFragment preserving the back stack - clicking on the back
button from the new MapFragment would navigate the user back
to the previous secondary screen.

One consequence of this change is that now it's possible to
get multiple instances of MapFragment on the back stack.
While testing I couldn't find any issue with that and assume
that it's not an issue.