Close the topo bottom sheet on back press
Prior to this change pressing the back button on the
topo bottom sheet would close the app (or navigate back
in the back stack).
This was a bit annoying as the standard UX behavior
on Android should be closing the bottom sheet on back press
click.

Fix that by implementing a custom OnBackPressedCallback
which is enabled while a topo is opened.

I couldn't find a way of unselecting the problem in the
view model and instead exposed the existing
BoulderMap#unselectProblem method. I also added notifyListener
parameter to it to automatically trigger the listener (if any).
IMO it would be better to flip the default flag and set it
to `true` and only use `false` where justified (I suppose
in some cases we don't want to run bottom sheet toggle animations
too often when, for example, selecting a new topo).